make _PyGenObject_HEAD private, etc
